+++ TM4C123_ADC/xDriver_MCU/GPIO/Driver/Intrinsics/Interrupt/InterruptRoutine/xSource/GPIO_InterruptRoutine_Vector_GPIOC.c
@@ -47,7 +47,7 @@ void GPIOC__vIRQVectorHandler(void)
          }
      }
 
-     if(GPIO_enBUS_APB ==u32RegBUS)
+     if((uint32_t)GPIO_enBUS_APB ==u32RegBUS)
      {
          u32Reg=GPIOC_APB_GPIOMIS_R;
          if(u32Reg & GPIO_enPIN_0)
